core: move formatting of plot info to string-format.cpp
Arguably, string handling in C is painful. Move the function
that formats the "tooltip" of the current spot on the profile
to the C++ string-format.cpp.

Note: the actual code stays C with a few necessary adaptions
concerning the translation machinery.
